vivify-backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic==2.5.0
google-cloud-monitoring==2.18.0
google-cloud-billing==1.12.0
google-cloud-compute==1.15.0
google-cloud-storage==2.14.0
google-cloud-container==2.38.0
google-cloud-resource-manager==1.15.0
google-auth==2.27.0
websockets==12.0

# Agent dependencies
langchain==0.2.16
langchain-community==0.2.16
google-generativeai>=0.3.0
tavily-python==0.3.0
sqlalchemy==2.0.23
aiosqlite==0.19.0

# Database
asyncpg==0.29.0
psycopg2-binary==2.9.9

# AWS (LocalStack)
boto3==1.34.0
botocore==1.34.0

# Terraform
python-hcl2==4.3.0

# Orchestration
# langgraph - using custom orchestrator instead

# Experiment utilities
pandas==2.2.0
